requirements.txt
google-generativeai
python-redmine
python-dotenv
requests
pytest
pytest-mock
